added bee movement  (using bezier curves)

+++ Minigin/BezierPath.h
@@ -38,7 +38,7 @@ public:
 
 	void Sample(std::vector<glm::vec2>* sampledPath)
 	{
-		for (size_t i = 0; i < m_BezierCurves.size(); i++)
+		for (int i = 0; i < m_BezierCurves.size(); i++)
 		{
 			for (float t = 0; t <= 1.0f; t += (1.0f / m_Samples[i]))
 			{
